+++ include/ortographicCamera.hpp
@@ -13,17 +13,17 @@ public:
     {
       if (arg.getKey() == "screen_window")
       {
-        screenWindow = arg.getValues<float>();
+        screenWindow = arg.getValues<double>();
       }
     }
   };
 
-    Ray generate_ray(int x, int y) const
-    {
-      float u = screenWindow[0] + (x + 0.5) * (screenWindow[1] - screenWindow[0]) / film->getWidth();
-      float v = screenWindow[2] + (y + 0.5) * (screenWindow[3] - screenWindow[2]) / film->getHeight();
-      return Ray(this->e + u * this->u + v * this->v, this->w);
-    }
+  Ray generate_ray(int x, int y) const
+  {
+    double u = screenWindow[0] + (x + 0.5) * (screenWindow[1] - screenWindow[0]) / film->getWidth();
+    double v = screenWindow[2] + (y + 0.5) * (screenWindow[3] - screenWindow[2]) / film->getHeight();
+    return Ray(this->e + u * this->u + v * this->v, this->w);
+  }
 
   static Camera *Make(Arguments args)
   {
@@ -44,10 +44,10 @@ public:
     if (screenWindow.empty())
     {
 
-      float xratio, yratio;
-      yratio = std::max(film->getHeight(), film->getWidth()) / float(film->getWidth());
-      xratio = std::max(film->getHeight(), film->getWidth()) / float(film->getHeight());
-      screenWindow = std::vector<float>{-xratio, xratio, -yratio, yratio};
+      double xratio, yratio;
+      yratio = std::max(film->getHeight(), film->getWidth()) / double(film->getWidth());
+      xratio = std::max(film->getHeight(), film->getWidth()) / double(film->getHeight());
+      screenWindow = std::vector<double>{-xratio, xratio, -yratio, yratio};
     }
     //    std::cout << "common setFilm" << std::endl;
   }
